Added more functionality to the database.
Now when we get the JSON data, we parse it, and extract the single
news items.

After that we save single news items in the database,
paying attention, not to save the same news twice. For example when
we do the sync multipy times really fast, we could get the same news,
so we put unique constrainte on two filed in the database table,
image url and publication data, and in that way we will have only
single news in the database.