Recycler View implementation
Added the Recycler View, and appropriate adapter that extends it.
Updated the Loader to work with new adapter.

Added the appropriate onItemClick listener to ViewHolder. It now
informes the NewsFeedListActivity when there is a onClick of some
of the items in the list, after which NewsFeedListActivity display
detailed information about the news.

Small changes when saving the files.