Adding the Search View and Material Toolbar
Both activities have the Toolbar as the ActionBar. First activity,
includes the menu in actionbar, and the second has HomeAsUp button
enabled.

SearchView added to the list activity, and logic for the
searching is now working. There is still things left to do.

Loader changed to support searching of the data based on text
that user enteres in the SearchView. We only search for the words
in the Title of the news.